# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FV3M",
        "contact": {
            "email": "mbel002@mbel.co.jp"
        },
        "license": {
            "name": "MBEL",
            "url": "http://www.mbel.co.jp/"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://3.112.130.195:8080/api/v1",
            "description": "Staging API Server"
        },
        {
            "url": "http://fv3m.jp:8000/api/v1",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "operationId": "App\\Http\\Controllers\\API\\UserController::index",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "path",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "returns list of users with pagination .",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": ""
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied!."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Store new user",
                "description": "Returns user data",
                "operationId": "App\\Http\\Controllers\\API\\UserController::store",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "role",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ponpeera"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mbel001@mbel.co.jp"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "imageUrls": {
                                        "type": "string",
                                        "example": "https://5.imimg.com/data5/AL/CC/MY-19161367/counterbalanced-forklift-250x250.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "returns stored user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied!"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user By Id",
                "description": "Get Individual user data according to user-id",
                "operationId": "App\\Http\\Controllers\\API\\UserController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "returns user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user",
                "description": "updates user data",
                "operationId": "App\\Http\\Controllers\\API\\UserController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "role"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "ponpeera"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mbel001@mbel.co.jp"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "imageUrls": {
                                        "type": "string",
                                        "example": "https://5.imimg.com/data5/AL/CC/MY-19161367/counterbalanced-forklift-250x250.png"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "returns updated user data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": ""
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied!"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}